app-cp/requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysql-connector-python==8.0.15
protobuf==3.6.1
redis==3.2.0
six==1.12.0
Werkzeug==0.14.1
